# Lab2/openapi.yaml
openapi: 3.0.3
info:
  title: Computer Store
  version: 1.0.0
...
paths:
...
  /orders/{id}:
    get:
      tags:
        - orders
      summary: Get orders by id
      description: Get orders by id
      operationId: GetordersById
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/orders'
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/orders'
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
    put:
      tags:
        - orders
      summary: Updates status orders
      description: ''
      operationId: updateorders
      parameters:
        - name: id
          in: path
          description: ID of products that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/orders'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: unautorized
        '403':
          description: forbidden
        '404':
          description: Not found
        '500':
          description: Status internal server error
    delete:
      tags:
        - orders
      summary: Deletes a orders
      description: delete a orders
      operationId: deleteorders
      parameters:
        - name: id
          in: path
          description: users id to orders
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
...
  /itemorders:
    post:
      tags:
        - itemorders
      summary: Add a new itemorders
      description: Add a new itemorders
      operationId: addItemorders
      requestBody:
        description: Create a new item orders
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/itemorders'
          application/xml:
            schema:
              $ref: '#/components/schemas/itemorders'
        required: true
      responses:
        '201':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
  /itemorders/{id}:
    get:
      tags:
        - itemorders
      summary: Get itemorders by id
      description: Get itemorders by id
      operationId: GetItemordersById
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/itemorders'
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/itemorders'
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
    put:
      tags:
        - itemorders
      summary: Updates itemorders
      description: ''
      operationId: updateItemorders
      parameters:
        - name: id
          in: path
          description: ID of products that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/itemorders'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: unautorized
        '403':
          description: forbidden
        '404':
          description: Not found
        '500':
          description: Status internal server error
    delete:
      tags:
        - itemorders
      summary: Deletes a itemorders
      description: delete a itemorders
      operationId: deleteItemorders
      parameters:
        - name: id
          in: path
          description: users id to itemorders
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
  /itemorders?orders_id=int:
    get:
      tags:
        - itemorders
      summary: Get itemorderss by orders_id
      description: Get itemorderss by orders_id
      operationId: GetItemordersByIdorders
      parameters:
        - name: orders_id
          in: query
          description: The id that needs to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/itemorders'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/itemorders'
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
  /carts:
    post:
      tags:
        - carts
      summary: Add a new carts
      description: Add a new carts
      operationId: addcarts
      requestBody:
        description: Create a new carts
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/carts'
          application/xml:
            schema:
              $ref: '#/components/schemas/carts'
        required: true
      responses:
        '201':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
  /carts? users_id = int:
    get:
      tags:
        - carts
      summary: Get carts by users_id
      description: Get carts by users_id
      operationId: GetcartsByIdusers
      parameters:
        - name: users_id
          in: query
          description: The id that needs to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/carts'
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/carts'
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
  /carts/{id}:
    get:
      tags:
        - carts
      summary: Get carts by id
      description: Get carts by id
      operationId: GetcartsById
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/carts'
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/carts'
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
    put:
      tags:
        - carts
      summary: Updates carts
      description: ''
      operationId: updatecarts
      parameters:
        - name: id
          in: path
          description: ID of products that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/carts'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: unautorized
        '403':
          description: forbidden
        '404':
          description: Not found
        '500':
          description: Status internal server error
    delete:
      tags:
        - carts
      summary: Deletes a carts
      description: delete a carts
      operationId: deletecarts
      parameters:
        - name: id
          in: path
          description: users id to carts
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
  /itemcarts:
    post:
      tags:
        - itemcarts
      summary: Add a new item carts
      description: Add a new itemcarts
      operationId: addItemcarts
      requestBody:
        description: Create a new item carts
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/itemcarts'
          application/xml:
            schema:
              $ref: '#/components/schemas/itemcarts'
        required: true
      responses:
        '201':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
  /itemcarts/{id}:
    get:
      tags:
        - itemcarts
      summary: Get itemcarts by id
      description: Get itemcarts by id
      operationId: GetItemcartsById
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/itemcarts'
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/itemcarts'
          description: Successful operation
        '400':
          description: Invalid input
        '500':
          description: Status internal server error
    put:
      tags:
        - itemcarts
      summary: Updates itemcarts
      description: ''
      operationId: updateItemcarts
      parameters:
        - name: id
          in: path
          description: ID of Itemcarts that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/itemcarts'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '401':
          description: unautorized
        '403':
          description: forbidden
        '404':
          description: Not found
        '500':
          description: Status internal server error
    delete:
      tags:
        - itemcarts
      summary: Deletes a itemcarts
      description: delete a itemcarts
      operationId: deleteItemcarts
      parameters:
        - name: id
          in: path
          description: users id to itemcarts
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: unautorized
        '403':
          description: forbidden
        '500':
          description: Status internal server error
components:
  schemas:
...
    carts:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        users_id:
          type: integer
          format: int64
          example: 1
        data_created:
          type: string
          format: date-time
          example: '2024-09-20'
      xml:
        name: carts
    itemcarts:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        carts_id:
          type: integer
          format: int64
          example: 1
        products_id:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          format: int64
          example: 1
      xml:
        name: itemcarts
    orders:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        users_id:
          type: integer
          format: int64
          example: 1
        status:
          type: string
          enum:
            - pending
            - delivering
            - delivered
          example: Delivered
        data_created:
          type: string
          format: date-time
          example: '2024-09-20'
      xml:
        name: orders
    itemorders:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        orders_id:
          type: integer
          format: int64
          example: 1
        products_id:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          format: int64
          example: 1
      xml:
        name: itemorders
...